requirements.txt
Django==5.0.6
djangorestframework==3.15.1
celery==5.3.6
redis==5.0.1
